Change the way of rendering the editor content
- Remove Sidebar component
- Implement renderView and renderEdit methods
